Starter/requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-allauth==0.54.0
django-crispy-forms==2.3
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-htmx==1.18.0
django-import-export==4.1.1
django-modeltranslation==0.19.5
django-template-partials==24.2
django-unfold==0.33.0
django-widget-tweaks==1.4.12
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==18.6.0
idna==3.4
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.5.0
pycparser==2.21
PyJWT==2.6.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
